fix(test): rename operator helpers to testclient getters
